requirements.txt
pandas==2.2.2
altair[all]==5.4.0
ipinfo==5.0.1
python-dotenv==1.0.1 
alive-progress==3.1.5 
scikit-learn
pylatex
